Relocate, simplify and rename pltMLMGResidual
Moved the plotting function to PeleLMeX_Plot.cpp and made the function a member of PeleLM like the other plotting functions. The function was renamed for consistency. Also, made use of GetVecOfPtrs and GetVecOfConstPtrs in the MAC projection.
