+++ Source/PeleLMeX_Utils.H
@@ -1,10 +1,6 @@
 #ifndef PELELM_UTILS_H
 #define PELELM_UTILS_H
 
-#include <AMReX_MLMG.H>
-#include <AMReX_MultiFab.H>
-#include <AMReX_Geometry.H>
-
 template <class T>
 amrex::Gpu::DeviceVector<T>
 convertToDeviceVector(const amrex::Vector<T>& v)
@@ -20,15 +16,5 @@ convertToDeviceVector(const amrex::Vector<T>& v)
 #endif
   return v_d;
 }
-
-void pltMLMGResidual(
-  amrex::MLMG& a_mlmg,
-  const amrex::Vector<amrex::MultiFab*>& a_sol,
-  const amrex::Vector<const amrex::MultiFab*>& a_rhs,
-  const std::string& a_solver_name,
-  const std::string& a_plot_file_root,
-  int a_step,
-  const amrex::Vector<amrex::Geometry>& a_geom,
-  class PeleLM* a_pelelm);
 
 #endif
